% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{add_prediction}
\alias{add_prediction}
\title{Add predictions and confidence interval to data using either statistical model or model coefficients}
\usage{
add_prediction(
  data,
  model = NULL,
  coefficients = NULL,
  coeff_cols = NULL,
  interval = c("none", "confidence", "prediction"),
  conf.level = 0.95
)
}
\arguments{
\item{data}{A data-frame containing appropriate values for all the terms in the model}

\item{model}{A regression model object which will be used to make predictions for the observations in `data`.
Will override `coefficients` if specified.}

\item{coefficients}{If a regression model is not available (or can't be fit in R), the regression coefficients from a model fit in some other language can be used to calculate predictions.
However, the user would have to ensure there's an appropriate one-to-one positional mapping between the data columns and the coefficient values.
Further, it would not be possible to calculate confidence/prediction intervals using this method.}

\item{coeff_cols}{If `coefficients` are specified and there isn't a one-to-one positional mapping between}

\item{interval}{Type of interval to calculate:
\describe{
  \item{"none" (default)}{No interval to be calculated.}
  \item{"confidence"}{Calculate a confidence interval.}
  \item{"prediction"}{Calculate a prediction interval.}
}}

\item{conf.level}{The confidence level for calculating confidence/prediction intervals. Default is 0.95.}
}
\value{
A data-frame with the following additional columns
 \describe{
   \item{.Pred}{The predicted response for each observation.}
   \item{.Lower}{The lower limit of the confidence/prediction interval (if possible) for each observation.}
   \item{.Upper}{The lower limit of the confidence/prediction interval (if possible) for each observation.}
 }
}
\description{
Add predictions and confidence interval to data using either statistical model or model coefficients
}
\examples{
library(DImodels)
data(sim1)

# Fit DI models with different ID effect groupings
mod <- lm(response ~ 0 + p1 + p2 + p3 + p4 + p1:p2 + p3:p4, data = sim1)

# Create new data for adding predictions
newdata <- head(sim1[sim1$block == 1,])
print(newdata)

# Adding predictions to data with confidence interval
add_prediction(data = newdata, model = mod)

# Calculate prediction intervals instead
add_prediction(data = newdata, model = mod, interval = "prediction")

# Change confidence level
add_prediction(data = newdata, model = mod, interval = "prediction",
               conf.level = 0.99)

# Use model coefficients for prediction
coeffs <- mod$coefficients

# Would now have to add columns corresponding to each coefficient in the
# data and ensure there is a desired mapping between data columns and
# the coefficients. Also confidence intervals cannot be calculated.
newdata$`p1:p2` = newdata$p1 * newdata$p2
newdata$`p3:p4` = newdata$p3 * newdata$p4

# If the coefficients are named then the function will try to
# perform matching between data columns and the coefficients
add_prediction(data = newdata, coefficients = coeffs)

# However, if the coefficients are not named
coeffs <- unname(coeffs)

# The user would have to manually specify the subset
# of data columnsvarranged according to the coefficients
subset_data <- newdata[, c(3:6, 8,9)]
add_prediction(data = subset_data, coefficients = coeffs)

# Or specify a selection (either by name or index) in coeff_cols
add_prediction(data = newdata, coefficients = coeffs,
               coeff_cols = c("p1", "p2", "p3", "p4", "p1:p2", "p3:p4"))

}
