% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionalTernary.R
\name{conditional_ternary_plot}
\alias{conditional_ternary_plot}
\title{Conditional ternary diagrams}
\usage{
conditional_ternary_plot(
  data,
  nlevels = 7,
  colours = NULL,
  lower_lim = NULL,
  upper_lim = NULL,
  tern_labels = c("P1", "P2", "P3"),
  contour_text = TRUE,
  show_axis_labels = TRUE,
  show_axis_guides = FALSE,
  axis_label_size = 4,
  vertex_label_size = 5
)
}
\arguments{
\item{data}{A data-frame which is the output of the
`\link{conditional_ternary_data}` function.}

\item{nlevels}{The number of levels to show on the contour map.}

\item{colours}{A character vector or function specifying the colours for the
contour map. The number of colours should be same as `nlevels`.
The default colours scheme is the
\code{\link[grDevices:terrain.colors]{terrain.colors()}}.}

\item{lower_lim}{A number to set a custom lower limit for the contour.
The default is minimum of the prediction.}

\item{upper_lim}{A number to set a custom upper limit for the contour.
The default is maximum of the prediction.}

\item{tern_labels}{A character vector containing the labels of the vertices
of the ternary. The default is "P1", "P2", and "P3".}

\item{contour_text}{A boolean value indicating whether to include labels on
the contour lines showing their values. The default is
\code{TRUE}.}

\item{show_axis_labels}{A boolean value indicating whether to show axis
labels along the edges of the ternary. The default
is \code{TRUE}.}

\item{show_axis_guides}{A boolean value indicating whether to show axis
guides within the interior of the ternary. The
default is \code{FALSE}.}

\item{axis_label_size}{A numeric value to adjust the size of the axis labels
in the ternary plot. The default size is 4.}

\item{vertex_label_size}{A numeric value to adjust the size of the vertex
labels in the ternary plot. The default size is 5.}
}
\value{
A data-frame with the following columns and any additional columns
        specified in `exp_str` parameter
 \describe{
   \item{.x}{The x component of the x-y projection of the simplex point.}
   \item{.y}{The y component of the x-y projection of the simplex point.}
   \item{.P1}{The first entity whose proportion is varied across the simplex.}
   \item{.P2}{The second entity whose proportion is varied across the simplex.}
   \item{.P3}{The third entity whose proportion is varied across the simplex.}
   \item{.add_str_ID}{An identifier column for grouping the cartesian product
                      of all additional columns specified in `exp_str`
                      parameter (if `exp_str` is specified).}
   \item{.Pred}{The predicted repsonse for each observation
               (if `prediction` is \code{TRUE}).}
   \item{.Lower}{The lower limit of the prediction/confidence interval
                 for each observation.}
   \item{.Upper}{The upper limit of the prediction/confidence interval
                 for each observation.}
 }

A ggplot object
}
\description{
The helper function for plotting conditional ternary diagrams. The output of
the `\code{\link{conditional_ternary_data}}` should be passed here to
visualise the n-dimensional simplex space as 2-d slices showing the change
in the response across any three variables, when a fourth is fixed at a
particular value.
}
\examples{
library(DImodels)

## Load data
data(sim2)

## Fit model
mod <- glm(response ~ 0 + (p1 + p2 + p3 + p4)^2, data = sim2)

## Create data for slicing
plot_data <- conditional_ternary_data(prop = c("p1", "p2", "p3", "p4"),
                                      conditional = "p4",
                                      model = mod,
                                      resolution = 1)

## Create plot
conditional_ternary_plot(data = plot_data)
}
