% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectsPlots.R
\name{effects_plot}
\alias{effects_plot}
\title{Visualise effect of species loss or addition}
\usage{
effects_plot(
  model,
  communities = NULL,
  species_interest = NULL,
  effect = c("addition", "loss"),
  .equi = FALSE,
  .design = TRUE,
  .se = FALSE,
  .average = TRUE,
  FG = NULL
)
}
\arguments{
\item{model}{A Diversity Interactions model object fit by using the \code{\link[DImodels:DI]{DI()}} function from the \code{\link[DImodels:DImodels-package]{DImodels}} package.}

\item{communities}{A dataframe specifying communities of interest for which user wants visualise the effect of species loss or addition. If left blank, the communities from the original data used to fit the model would be selected}

\item{species_interest}{A character vector specifying the species for which to visualise the effect of addition or loss on the response. If left blank, depending on the number of communities in the design, either the original communities or the equi-proportional community would be selected.}

\item{effect}{A character string with one value from c("addition", "loss" ) whether to visualise the effect of addition or loss of a species.}

\item{.equi}{A boolean variable indicating whether to include the equi-proportional community in the plot}

\item{.design}{A boolean variable to indicate whether to include the communities in the design in the plot}

\item{.se}{A boolean variable indicating whether to plot confidence intervals associated with the effect of species addition or loss}

\item{.average}{A boolean variable indicating whether to add a line describing the "average" effect of species addtion or loss}

\item{FG}{A character vector specifying the functional grouping of the species in the design. Species belonging to the same functional group will be assigned with different shades of the same colour. The user can manually specify a character vector giving the functional group each species belongs to. If left empty the functional will try to get a functional grouping from the original \code{\link[DImodels:DI]{DI}} model object.}
}
\value{
A ggplot object
}
\description{
This function helps to visualise the effect of the addition or loss of a species from a community on the response
}
\examples{
library(DImodels)

## Load data
data(sim1)

## Fit model
mod <- DI(prop = 3:6, DImodel = "AV", data = sim1, y = "response")

## Get effects plot for all species in design
effects_plot(model = mod)

## Add custom communities to plot instead of design communities
effects_plot(model = mod, .design = FALSE,
             communities = data.frame(p1 = c(0.9, 0.1),
                                      p2 = c(0.1, 0.9)))

## Add uncertainity on plot
effects_plot(model = mod,
             communities = data.frame(p1 = c(0.9, 0.1),
                                      p2 = c(0.1, 0.9)),
             .se= TRUE)

## Visualise effect of species loss for particular species
effects_plot(model = mod, effect = "loss",
             species_interest = c("p1", "p3"))
}
