% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionalTernary.R
\name{functional_ternary}
\alias{functional_ternary}
\title{Conditional ternary diagrams at functional group level}
\usage{
functional_ternary(
  model,
  FG,
  values = NULL,
  cond_values = NULL,
  cond_FG = NULL,
  resolution = 3,
  nlevels = 7,
  colours = NULL,
  lower_lim = NULL,
  upper_lim = NULL,
  .contour_text = T,
  .axis_guides = F,
  axis_label_size = 4,
  vertex_label_size = 5
)
}
\arguments{
\item{model}{A Diversity Interactions model object fit by using the \code{\link[DImodels:DI]{DI()}} function from the \code{\link[DImodels:DImodels-package]{DImodels}} package.}

\item{FG}{A character vector specifying the functional grouping of the species in the design. If left empty the functional will try to get a functional grouping from the original \code{\link[DImodels:DI]{DI}} model object.}

\item{values}{A vector specifying the proportional split between the species within a functional group. The default is to split the functional group proportional equally between each species.}

\item{cond_values}{A vector of numbers between 0 and 1 describing the values at which to takes slices of the ternary. Necessary only when there are more than 3 funcitonal groups.}

\item{cond_FG}{A character vector specifying the functional groups to slice the ternary on if there are more than three functional groups}

\item{resolution}{A number between 1 and 5 describing the resolution of the resultant graph. A high value would result in a higher definition figure but at the cost of being computationlly expernsive.}

\item{nlevels}{The number of levels for the contour map.}

\item{colours}{The colours for the contour map. The default colours scheme is the terrain.colors from hcl.pals()}

\item{lower_lim}{A number to set a custom lower limit for the contour. The default is minimum of the prediction}

\item{upper_lim}{A number to set a custom upper limit for the contour. The default is maximum of the prediction}

\item{.contour_text}{A boolean value indicating whether to include labels on the contour lines showing their values}

\item{.axis_guides}{A boolean value indicating whether to show axis guides in the ternary}

\item{axis_label_size}{A numeric value to adjust the size of the axis labels in the contour}

\item{vertex_label_size}{A numeric value to adjust the size of the vertex labels in the contour}
}
\value{
A ggplot object
}
\description{
Conditional ternary diagrams at functional group level
}
