% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradientChange.R
\name{gradient_change_data}
\alias{gradient_change_data}
\title{Visualising average change in response over a diversity gradient.}
\usage{
gradient_change_data(data, prop, gradient = c("richness", "evenness"), ...)
}
\arguments{
\item{data}{A data-frame consisting of species proportions and
any other necessary variables to make predictions from
`model` or `coefficients`.}

\item{prop}{A vector identifying the column-names or indices of the
columns containing the proportions in `data`.}

\item{gradient}{Diversity gradient to show on the X-axis, one of
"richness" or "evenness". Defaults to "richness".}

\item{...}{
  Arguments passed on to \code{\link[=add_prediction]{add_prediction}}
  \describe{
    \item{\code{model}}{A regression model object which will be used to make predictions for the observations in `data`.
Will override `coefficients` if specified.}
    \item{\code{coefficients}}{If a regression model is not available (or can't be fit in R), the regression coefficients from a model fit in some other language can be used to calculate predictions.
However, the user would have to ensure there's an appropriate one-to-one positional mapping between the data columns and the coefficient values.
Further, it would not be possible to calculate confidence/prediction intervals using this method.}
    \item{\code{coeff_cols}}{If `coefficients` are specified and there isn't a one-to-one positional mapping between}
    \item{\code{conf.level}}{The confidence level for calculating confidence/prediction intervals. Default is 0.95.}
    \item{\code{interval}}{Type of interval to calculate:
\describe{
  \item{"none" (default)}{No interval to be calculated.}
  \item{"confidence"}{Calculate a confidence interval.}
  \item{"prediction"}{Calculate a prediction interval.}
}}
  }}
}
\value{
A data-frame with the following columns appended at the end
 \describe{
   \item{.Pred}{The predicted response for each community.}
   \item{.Lower}{The lower limit of the prediction/confidence interval
                 for each observation.}
   \item{.Upper}{The upper limit of the prediction/confidence interval
                 for each observation.}
   \item{.Richness}{The richness (number of species) within each community.}
   \item{.Evenness}{The evenness (metric quantifying the relative abundance
                    of each species) of each community.}
   \item{.Gradient}{An character string defining the diversity gradient used
                    for averaging the response.}
   \item{.Avg}{The averaged value of the response for each unique value of
               the selected diversity gradient.}
 }
}
\description{
Helper function for creating the data for the average richness plot.
}
\examples{
library(DImodels)

## Load data
data(sim2)

## Fit model
mod <- glm(response ~ 0 + (p1 + p2 + p3 + p4)^2, data = sim2)

## Create data
## By default response would be averaged on the basis of richness
head(gradient_change_data(data = sim2,
                          prop = c("p1", "p2", "p3", "p4"),
                          model = mod))

## Average response with respect to evenness
head(gradient_change_data(data = sim2,
                          prop = c("p1", "p2", "p3", "p4"),
                          model = mod,
                          gradient = "evenness"))
}
