% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSelection.R
\name{model_selection}
\alias{model_selection}
\title{Visualising model selection}
\usage{
model_selection(
  models,
  metric = c("AIC", "BIC", "AICc", "BICc", "logLik"),
  sort = FALSE,
  model_names = names(models)
)
}
\arguments{
\item{models}{List of Diversity Interaction (DI) models fit by using the \code{\link[DImodels:DI]{DI()}} function from the \code{\link[DImodels:DImodels-package]{DImodels}} package.}

\item{metric}{Metric used for comparisons between models. Takes values from c("AIC", "BIC", "AICc", "BICc", "logLik"), Can choose a single or multiple metrics for comparing the different models}

\item{sort}{A boolean value indicating whether to sort the model from highest to lowest value of chosen metric}

\item{model_names}{A character string describing the names to display on X-axis for each model in order they appear in the models parameter}
}
\value{
A ggplot object
}
\description{
This function helps to visualise model selection by showing a visual comparison between models using information criteria
}
\examples{
library(DImodels)

## Load data
data(sim2)

## Fit different DI models
mod_AV <- DI(prop = 3:6, DImodel = "AV", data = sim2, y = "response")
mod_FULL <- DI(prop = 3:6, DImodel = "FULL", data = sim2, y = "response")
mod_FG <- DI(prop = 3:6, DImodel = "FG", FG = c("G","G","L","L"),
             data = sim2, y = "response")
mod_AV_theta <- DI(prop = 3:6, DImodel = "AV", data = sim2,
                   y = "response", estimate_theta = TRUE)
mod_FULL_theta <- DI(prop = 3:6, DImodel = "FULL", data = sim2,
                     y = "response", estimate_theta = TRUE)
mod_FG_theta <- DI(prop = 3:6, DImodel = "FG", FG = c("G","G","L","L"),
                   data = sim2, y = "response", estimate_theta = TRUE)

models_list <- list(mod_AV, mod_FULL, mod_FG,
                    mod_AV_theta, mod_FULL_theta, mod_FG_theta)

## Compare different information criteria for the different models
model_selection(models = models_list,
                metric = c("AIC", "AICc"))

## If single metric is specified then breakup of metric
## between loglik and penalty will be shown
model_selection(models = models_list,
                metric = c("AIC"))

## Sort models
model_selection(models = models_list,
                metric = c("AIC"), sort = TRUE)

## If multiple metrics are specified then sorting will be done on first metric
model_selection(models = models_list,
                metric = c("AIC", "logLik", "BIC"), sort = TRUE)

## By default the labels on the X-axis for the models will be
## created by assigning a unique ID to each model sequentially
## in the order they appear in the models object
model_selection(models = models_list,
                metric = c("AIC", "logLik", "BIC"), sort = TRUE)

## When possible the variables names of objects contatining the
## individual models would be used as axis labels
model_selection(models = list(mod_AV, mod_FULL, mod_FG,
                              mod_AV_theta, mod_FULL_theta, mod_FG_theta),
                metric = c("AIC", "logLik", "BIC"), sort = TRUE)

## If neither of these two situations are desirable custom labels
## for each model can be specified using the model_names parameter
model_selection(models = list(mod_AV, mod_FULL, mod_FG,
                              mod_AV_theta, mod_FULL_theta, mod_FG_theta),
                metric = c("AIC", "logLik", "BIC"), sort = TRUE,
                model_names = c("AV model", "Full model", "FG model",
                                "AV theta", "Full theta", "FG theta"))
}
