% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualiseEffects.R
\name{visualise_effects}
\alias{visualise_effects}
\title{Visualise effect of species addition or loss}
\usage{
visualise_effects(
  model,
  data = NULL,
  species_interest = NULL,
  effect = c("addition", "loss"),
  exp_str = list(),
  equi = FALSE,
  design = TRUE,
  se = FALSE,
  average = TRUE,
  FG = NULL,
  conf.level = 0.95,
  nrow = 0,
  ncol = 0
)
}
\arguments{
\item{model}{A Diversity Interactions model object fit by using the
`\code{\link[DImodels:DI]{DI()}}` function from the
`\code{\link[DImodels:DImodels-package]{DImodels}}` package.}

\item{data}{A dataframe specifying communities of interest for which user
wants visualise the effect of species loss or addition.
If left blank, the communities from the original data used
to fit the model would be selected.}

\item{species_interest}{A character vector specifying the species for which to visualise
the effect of addition or loss on the response. If left blank,
all species would be assumed to be of interest.}

\item{effect}{A character string with one value from c("addition", "loss") whether to
visualise the effect of addition or loss of a species.}

\item{exp_str}{A list specifying values for additional experimental
structures in the model other than the proportions.
This would be useful to compare the predictions across
different values for a categorical variable. One plot will
be generated for each unique combination of values specified
here and they will be arranged in a grid according to the
value specified in `nrow` and `ncol`.}

\item{equi}{A boolean variable indicating whether to include the
equi-proportional community in the plot.}

\item{design}{A boolean variable to indicate whether to include the
communities in the design in the plot.}

\item{se}{A boolean variable indicating whether to plot confidence intervals associated with
the effect of species addition or loss}

\item{average}{A boolean variable indicating whether to add a line describing the "average"
effect of species addition or loss}

\item{FG}{The functional grouping of the species in the design. Species
belonging to the same functional group will be assigned with
different shades of the same colour. The user can manually specify
a character vector giving the functional group each species
belongs to. If left empty the function will try to get a functional
grouping from the original \code{\link[DImodels:DI]{DI}} model
object.}

\item{conf.level}{The confidence level for calculating confidence or
prediction intervals.}

\item{nrow}{Number of rows in which to arrange the final plot
(when `exp_str` is specified).}

\item{ncol}{Number of columns in which to arrange the final plot
(when `exp_str` is specified).}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot}} ggplot or
        \code{\link[ggfortify:ggmultiplot-class]{ggfortify::ggmultiplot}}
        object.
}
\description{
This function helps to visualise the effect of the addition or loss
             of a species from a community on the response
}
\examples{
library(DImodels)

## Load data
data(sim1)

## Fit model
mod <- DI(prop = 3:6, DImodel = "AV", data = sim1, y = "response")

## Get effects plot for all species in design
visualise_effects(model = mod)

## Add custom communities to plot instead of design communities
visualise_effects(model = mod, design = FALSE,
                  data = data.frame("p1" = c(0.9, 0.1),
                                    "p2" = c(0.1, 0.9)))

## Add uncertainty on plot
visualise_effects(model = mod,
                  data = data.frame("p1" = c(0.9, 0.1),
                                    "p2" = c(0.1, 0.9)),
                  se= TRUE)

## Visualise effect of species loss for particular species
visualise_effects(model = mod, effect = "loss",
                  species_interest = c("p1", "p3"))
}
