% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{get_equi_comms}
\alias{get_equi_comms}
\title{Get all equi-proportional communities at each level of richness}
\usage{
get_equi_comms(
  nvars,
  variables = paste0("Var", 1:nvars),
  threshold = round(1e+06/nvars)
)
}
\arguments{
\item{nvars}{Number of variables in the design}

\item{variables}{Names for the variables. Will be used as column names for the
final result. Default is "Var" followed by column number.}

\item{threshold}{The maximum number of communities to select for each level
of richness for situations when there are too many
equi-proportional communities.}
}
\value{
A dataframe consisting all or a random selection of equi-proportional
        communities at each level of richness
}
\description{
Get all equi-proportional communities at each level of richness
}
\examples{
## Get all equi-proportional communities for each level of richnness upto 10
get_equi_comms(10)

## Change species names
get_equi_comms(4, variables = c("Lollium perenne", "Chicorum intybus",
                                "Trifolium repens", "Trifolium pratense"))
}
