% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{get_equi_comms}
\alias{get_equi_comms}
\title{Get all equi-proportional communities at each level of richness}
\usage{
get_equi_comms(
  nSpecies,
  species = paste0("Sp", 1:nSpecies),
  threshold = round(1e+06/nSpecies)
)
}
\arguments{
\item{nSpecies}{Number of species in the design}

\item{species}{Names for the species. Will be used as column names for the final result. Default is "Sp" followed by column number.}

\item{threshold}{The maximum number of communities to select for each level of richness for situations when there are too many equi-proportional communities.}
}
\value{
A dataframe consisting all or a random selection of equi-proportional communities at each level of richness
}
\description{
Get all equi-proportional communities at each level of richness
}
\examples{
## Get all equi-proportional communities for each level of richnness upto 10
get_equi_comms(10)

## Change species names
get_equi_comms(4, species = c("Lollium perenne", "Chicorum intybus",
                              "Trifolium repens", "Trifolium pratense"))
}
